Moved ItemCombination tests into new folder and separated the two files.
